requirements.txt
flask
openai
Pillow
torch
transformers
ultralytics
colorthief
opencv-python-headless
scikit-learn
requests
python-dotenv
fiftyone>=0.23.0
